feat(aggregate_tables): add options to show all counts and count threshold
Introduce `--keys-to-show-all-count` and `--show-count-threshold`
arguments to allow more flexible aggregation. This change enables
users to specify keys for which all counts should be displayed and
set a threshold for showing counts, enhancing the utility of the
aggregation tool.

refactor(aggregate): use ValueCounter class for counting values

Replace dictionary-based counting with a new `ValueCounter` class
to improve code readability and maintainability. This refactor
encapsulates counting logic within a dedicated class, making it
easier to manage and extend in the future.